backend/pyproject.toml
[tool.poetry]
name = "empathy-for-learning"
version = "0.1.0"
description = "UNESCO SEEK-based emotional support chatbot"
authors = ["Your Team <team@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
pymongo = "^4.5"
redis = "^5.0"
boto3 = "^1.28"
python-dotenv = "^1.0"
bcrypt = "^4.1"
pyjwt = "^2.8"
httpx = "^0.25"
email-validator = "^2.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
black = "^23.10"
flake8 = "^6.1"
mypy = "^1.6"
pytest-cov = "^4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
